Introduce the capability to "not use" the baseline from the PDE
This means one can re-use the API tooling from the PDE but can also
decide to not use it and only keep the semantic versioning builder.
(the PDE api tooling introduces quite a big overhead on build times)

This commit introduce a preference at the project level to specify 
where the baseline jars should be found.

User can either pick platform:/pde/apibaselines to look for jar using
the baselined declared in the PDE 

or a platform:/resource/someproject/someFile.target and the baseliner
will provision this target and look for the Jars there.